Move `create_sanitised_html_for_url` to `formatters.py`
It’s not specific to Markdown parsing because we also use it to preview
URLs in text messages, which don’t use Markdown
